+++ src/main/java/com/jvmup/nbbs/service/PostService.java
@@ -2,12 +2,16 @@ package com.jvmup.nbbs.service;
 
 import com.jvmup.nbbs.dao.PostDao;
 import com.jvmup.nbbs.dao.UserDao;
+import com.jvmup.nbbs.exception.DataInValidException;
+import com.jvmup.nbbs.exception.NoPermissionToGetData;
 import com.jvmup.nbbs.po.Post;
 import com.jvmup.nbbs.util.PageParam;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.jvmup.nbbs.po.User;
+
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -20,13 +24,24 @@ import java.util.List;
 @Service
 public class PostService {
     private PostDao postDao;
+    private UserDao userDao;
     public User user;
+    private UserService userService;
+
+    @Autowired
+    public void setUserService(UserService userService) {
+        this.userService = userService;
+    }
 
     @Autowired
     public void setPostDao(PostDao postDao) {
         this.postDao = postDao;
     }
 
+    @Autowired
+    public void setUserDao(UserDao userDao) {
+        this.userDao = userDao;
+    }
 
     /**
      * gzp 717 1550
@@ -34,17 +49,17 @@ public class PostService {
      * @param sectionId
      * @return
      */
-    public List<Post> getPostBySection(PageParam pageParam,  int sectionId, int type){
+    public List<PostWrapper> getPostBySection(PageParam pageParam,  int sectionId, int type,String nickname){
 
-        return postDao.getPost(pageParam,sectionId,type);
+        return changePostList(nickname,postDao.getPost(pageParam,sectionId,type));
     }
     /**
      * gzp 717 1553
      * 获取首页推荐的热帖
      * @return list
      */
-    public List<Post> getHotPost(){
-        return postDao.getHotPost();
+    public List<PostWrapper> getHotPost(String nickname){
+        return changePostList(nickname,postDao.getHotPost());
     }
     /**
      * 通过id获得post
@@ -58,22 +73,46 @@ public class PostService {
      * 删除post
      * @param id
      */
-    public void deletePost(int id){
-        postDao.deletePost(id);
+    public void deletePost(int id,int userId) throws Exception {
+        Integer a = userDao.getIdByPostId(id);
+        if (a==null)
+            throw new DataInValidException();
+        if (userDao.getIdByPostId(id)==userId)
+            postDao.deletePost(id);
+        else if (userDao.isAdmin(userId)==1||hasPermission(id,userId))
+            postDao.deletePost(id);
+        else
+            throw new NoPermissionToGetData("没有权限删除");
     }
     /**
      * 添加post
      * @param post
      */
    public void addPost(Post post){
+
         postDao.addPost(post);
+        userDao.updateEx(post.getUser().getId(),10);
     }
     /**
      * 置顶post
      * @param id
      */
-   public void toppingPost(int id){
-        postDao.toppingPost(id);
+   public void toppingPost(int id,int userId,int toppingType){
+       if (toppingType==2){
+           if (userDao.isAdmin(userId)==1){
+               postDao.toppingPost(id,toppingType);
+           }else{
+               throw new NoPermissionToGetData("没有权限去置顶该帖子");
+           }
+       }else{
+           if (hasPermission(id,userId)){
+               postDao.toppingPost(id,toppingType);
+           }else{
+               throw  new NoPermissionToGetData("没有权限去置顶该帖子");
+           }
+       }
+
+
     }
     /**
      * gzp
@@ -81,17 +120,20 @@ public class PostService {
      * @param postId
      *
      */
-    public void changeType(int postId,int sectionId){
+    public void changeType(int userId, int postId,int sectionId){
+        if (userDao.isAdmin(userId)==1||hasPermission(postId,userId))
         postDao.changeType(postId,sectionId);
+        else
+            throw new NoPermissionToGetData("没有权限去改变文章类型");
     }
     /**
      *
      * 点赞
-     * @param postId
      * @param userId
      */
-    public void praise(int postId,int userId){
-        postDao.praise(postId,userId);
+    public void praise(Post post,int userId){
+        postDao.praise(post.getId(),userId);
+        userDao.updateEx(userService.getIdByPostId(post.getId()),2);
     }
     /**
      * @param id postid
@@ -104,8 +146,8 @@ public class PostService {
      * 根据作者ID获得post
      * @param id 作者ID
      */
-    public List<Post> getPostByAuthId(int id){
-        return postDao.getPostByAuthId(id);
+    public List<PostWrapper> getPostByAuthId(int id,String nickname){
+        return changePostList(nickname,postDao.getPostByAuthId(id));
     }
     /**
      * 根据文章ID获取收藏次数
@@ -130,14 +172,6 @@ public class PostService {
     public void Report(int userId,int postId,String reason) {
         postDao.Report(userId, postId, reason);
     }
-    /**
-     * gzp 718 1303
-     *
-     */
-    public void getPostDetail(int postId){
-        postDao.getPraiseCnt(postId);
-        postDao.getCollectionCnt(postId);
-    }
 
     /**
      * 阅读次数
@@ -146,6 +180,85 @@ public class PostService {
 
     public void incView(int id){
         postDao.incView(id);
+    }
+
+    public List<PostWrapper> getToppingPost(String nickname){
+        return changePostList(nickname,postDao.getToppingPost());
+    }
+    private boolean hasPermission(int postId,int userId){
+        Post post = postDao.getPostByPostId(postId);
+        if (userDao.isAs(userId,post.getSection().getId())>=1){
+            return true;
+        }else
+            return userDao.isAp(userId, post.getSection().getPartition().getId()) >= 1;
+    }
+
+    static class PostWrapper{
+        Post post;
+        boolean like =false;
+        boolean collection=false;
+
+        public boolean isCollection() {
+            return collection;
+        }
+
+        public void setCollection(boolean collection) {
+            this.collection = collection;
+        }
+
+        public Post getPost() {
+            return post;
+        }
+
+        public void setPost(Post post) {
+            this.post = post;
+        }
+
+        public void setLike(boolean like) {
+            this.like = like;
+        }
+
+
+        public boolean isLike() {
+            return like;
+        }
+
+        public PostWrapper() {
+        }
+    }
+    private PostWrapper changePost(String nickname,Post post){
+        PostWrapper postWrapper = new PostWrapper();
+        postWrapper.setPost(post);
+        if (post.getLikeList().size()!=0)
+        for (User user:post.getLikeList()){
+            if (user.getNickname().equals(nickname)){
+                postWrapper.setLike(true);
+                break;
+            }
+        }
+        if (post.getCollectionList().size()!=0)
+        for(User user:post.getCollectionList()){
+            if (user.getNickname().equals(nickname)){
+                postWrapper.setCollection(true);
+                break;
+            }
+        }
+        return postWrapper;
+    }
+
+    private List<PostWrapper> changePostList(String nickname,List<Post> posts){
+        List<PostWrapper> wrappers = new ArrayList<>();
+        for (Post post:posts){
+            wrappers.add(changePost(nickname,post));
+        }
+        return wrappers;
+    }
+
+    public void cancelPraise(int userId,int postId){
+        postDao.cancelPraise(postId,userId);
+    }
+    public void cancelCollection(int userId,int postId){
+        postDao.cancelCollection(postId,userId);
     }
 }
